pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "assemblyai>=0.37.0",
    "fastapi>=0.115.11",
    "openai>=1.66.5",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "streamlit-webrtc>=0.53.3",
    "streamlit>=1.43.2",
]
